requirements.txt
asgiref==3.6.0
astroid==2.14.2
autopep8==2.0.1
boto3==1.28.77
botocore==1.31.77
certifi==2022.12.7
charset-normalizer==3.0.1
dill==0.3.6
Django==3.2
django-cors-headers==3.13.0
django-rest-framework==0.1.0
djangorestframework==3.14.0
docker==6.0.1
gunicorn==20.1.0
idna==3.4
isort==5.12.0
jmespath==1.0.1
lazy-object-proxy==1.9.0
mccabe==0.7.0
packaging==23.0
platformdirs==3.0.0
psycopg2-binary==2.9.5
pycodestyle==2.10.0
pylint==2.16.2
python-dateutil==2.8.2
python-decouple==3.8
python-dotenv==1.0.0
pytz==2022.7.1
requests==2.28.2
s3transfer==0.7.0
six==1.16.0
sqlparse==0.4.3
tomli==2.0.1
tomlkit==0.11.6
typing_extensions==4.5.0
urllib3==1.26.14
websocket-client==1.5.1
wrapt==1.14.1
